Create Services and endpoints (#2)
Authored-by: Elias Santos Garcia <elias.sgarcia@gmail.com>

+++ src/NutritionistPreview.Api.Core/Domain/Entities/Food.cs
@@ -6,6 +6,6 @@ namespace NutritionistPreview.Api.Core.Domain.Entities
     {
         public string Name { get; set; }
         public string Group { get; set; }
-        public double CaloricAmount { get; set; }
+        public int CaloricAmount { get; set; }
     }
 }
